refactor: rename calculateCommitteeAssignments.ts to shufflingUtils.ts
