+++ packages/state-transition/src/util/index.ts
@@ -5,7 +5,7 @@ export * from "./attesterStatus.js";
 export * from "./balance.js";
 export * from "./blindedBlock.js";
 export * from "./blockRoot.js";
-export * from "./calculateCommitteeAssignments.js";
+export * from "./shufflingUtils.js";
 export * from "./capella.js";
 export * from "./computeAnchorCheckpoint.js";
 export * from "./deposit.js";
